Split the row into a separate file
